Atualização 2.0

+++ Entities/Rpg/RpgMonsters/BaseMonstersEntity.cs
@@ -12,8 +12,8 @@ namespace FirstBotDiscord.Entities.Rpg.RpgMonsters
         public int MonsterId { get; set; }
         public string MonsterName { get; set; }
         public int Level { get; set; }
-        public int MonsterLife { get; set; }
-        public int MonsterMana { get; set; }
+        public double MonsterLife { get; set; }
+        public double MonsterMana { get; set; }
         public AtributesEntity MonsterAtributes { get; set; } = new AtributesEntity();
         public DateTime Spawn { get; set; }
     }
